+++ react/components/SearchQuery.js
@@ -169,6 +169,21 @@ const useCorrectSearchStateVariables = (
   return result
 }
 
+const sponsoredProductsResult = (
+  sponsoredProductsLoading,
+  sponsoredProductsError,
+  sponsoredProducts
+) => {
+  const showSponsoredProducts =
+    !sponsoredProductsLoading && !sponsoredProductsError
+
+  const sponsoredProductsResponse = showSponsoredProducts
+    ? sponsoredProducts
+    : []
+
+  return sponsoredProductsResponse
+}
+
 const useQueries = (
   variables,
   facetsArgs,
@@ -188,12 +203,19 @@ const useQueries = (
     },
   })
 
-  const { data: { sponsoredProducts } = [] } = useQuery(
-    sponsoredProductsQuery,
-    {
-      variables,
-      skip: sponsoredProductsBehavior === 'skip',
-    }
+  const {
+    data: { sponsoredProducts } = [],
+    loading: sponsoredProductLoading,
+    error: sponsoredProductsError,
+  } = useQuery(sponsoredProductsQuery, {
+    variables,
+    skip: sponsoredProductsBehavior === 'skip',
+  })
+
+  const sponsoredProductsReturn = sponsoredProductsResult(
+    sponsoredProductLoading,
+    sponsoredProductsError,
+    sponsoredProducts
   )
 
   const {
@@ -285,7 +307,7 @@ const useQueries = (
         sampling: facets?.sampling,
       },
       searchMetadata,
-      sponsoredProducts,
+      sponsoredProducts: sponsoredProductsReturn,
     },
     productSearchResult,
     refetch,